refactor(validation): extract validation result formatting
